Add files via upload
this is the first working version of roulette
it is extremely barebones and needs a lot added however, it is simply a safe save point because it simply works
